save data early in acoustic lab
ordered the save of the data before the waitforbuttonpress so that students dont accidentally stop the script before the data is saved. this may try to save the serial/oscope connection objects, but with warnings off this should not throw an error. could potentially exclude those object variables from the save function using a regex?